mmc: meson: Add driver for the SD/MMC host found on Amlogic MesonX SoCs
Add a driver for the SD/MMC host found on the Amlogic MesonX SoCs. It
is an MMC controller which provides an interface between the
application processor and various memory cards. It supports the SD
specification v2.0 and the eMMC specification v4.41. It also supports
SDSC, SDHC and SDXC memory card default speed.

This patch adds also the binding documentation.

Signed-off-by: Carlo Caione <carlo@endlessm.com>