adg: ensure we can unset SSICKR bit 31
The rsnd_mod_bset() call to update SSICKR in rsnd_adg_get_clkout()
can try to set bit 31, but is missing the bitmask to allow it to
be cleared.

On the R8A77795 SoC this seems to mean that clkout3 is still selected
from the A channel instead of B channel and therefore is operating at
the wrong rate.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>