pfc-r8a7795: fix issue with MSIOF0 only using GPSR5
The pfc-r8a7795 definitions do not work with devices that only
have an GPSR<x> control and no IPSR. This means that when selecting
a line such as MSIFO0_CLK, it fails to set the GPSR5 register and
returns an error.

Add a mapping for the function into the pinmux_data and also add
it to the GPSR5 array so the sh-pfc core can find these and set
them.

This should fix:
spi_sh_msiof e6e90000.spi: Error applying setting, reverse things back