i2c: i2c-rcar: enable pm before adding adapter
Before calling i2c_add_numbered_adapter() make sure the PM is
initialised by callling pm_runtime_enable() as the act of adding
the adapter may cause a device to probe that wants to make i2c
transfers at probe time.

This fixes the probe errors on the RCar-H3 when the PCA driver
is built into the kernel.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>